Use pattern to choose content model, unit tests
Instead of the optional sub-namespaces and other cruft,
each config now uses a 'pattern' param to choose which pages apply.
If there is no 'pattern', all pages in the given namespace are used.
If namespace is mentioned by some configs, but after trying them all,
the title did not match any patterns, the new page will not be allowed
to be created.

Behavior changes:
* isLocal is now true by default

Change-Id: I572299ddad8c5e6366d5eb0b9d708138c9fc1b19