Introduce basic edit dialog for editing tabular data
Add an "Edit as table" button on pages with 'Tabular.JsonConfig'
content, which opens a dialog for editing the data in a table widget.
On clicking "Apply", the raw JSON in the editing textbox is updated.

This takes advantage of the strict 'Tabular.JsonConfig' format to
find the data and column headings.

A feature flag, wgJsonConfigUseGUI, determines availability of the
feature.

For now, the dialog may not be used if the dataset is too large, or
if the JSON is not valid; in these cases, show an alert with an
error.

Bug: T251759
Change-Id: I5d6128c8af0b6734c81cc514647b4e498673171e