Add temporary compatibility before calling Kartographer
The actual issue is in Kartograper (which is called here).
Some of the code in SimpleStyleParser::normalizeAndSanitize()
expects arrays of objects. That's why this strange $dummy
thing is here. The problem is: Sometimes $data->data is
already an array of objects. In this case $dummy must not be
used but $data->data be passed as it is.

This is temporary and can be replaced with a trivial
`$ssp->normalizeAndSanitize( $data->data )` call when
Kartographer is fixed.

Bug: T281224
Change-Id: I2beced289c9b31ecb8d921c7906bfa9094f837d7