SECURITY: HTML-escape strings and don’t display arrays
The is_string() check is necessary to guard against arrays – they can’t
be stored, but the extension still tries to show them when previewing an
edit that tries to store an array column. Falling back to the empty
string seems good enough in that case (an error message is already shown
elsewhere).

includes a fix to also allow numbers

Bug: T302192
Change-Id: Ifa2f6ec9ccce3b781e83a9905392e70d6a7340ad