Use Codex for rest of the colors in JsonConfig
This replaces colors the used colors with similar ones
provided by Codex.

The only note about the change is use of @color-progressive
in .mw-jsonconfig-same and .mw-jsonconfig-value-info. Given
the same color didn't exist in Codex's palette this felt
to be most matching semantically though still is different
from what was before, #800080.

Change-Id: I88db2a143fed5145d40e75ee0491b252fc4d93a4