Make sure map's JSON at least has LTR direction
This won't happen in Wikimedia deployments where SyntaxHighlight
and Kartographer extensions exist but if the two don't exist and one wants to see a map in an RTL wiki which get a <pre>, this ensures
that at least has an LTR direction.

To test this create a RTL language wiki (e.g. Persian) with the
extension but don't enable Kartographer or SyntaxHighlight and create
a map data. The data should be in LTR to be readable just like when
SyntaxHighlight is enabled but before this change it is in RTL and
it isn't easy to follow.

Bug: T371853
Change-Id: Ic7666af26248a5f7694d2441a1e6347803cec70a