Fix exception display when JSON does not validate
When trying to preview a JSON map file that contains a JSON validation
error, $content->getLocalizedData() returns null, which can lead to the
ParserOutput still having a null mRawText, which in turns throws when it
is accessed. We do an explicit check of the null case here to take this
case into account, and set the raw text to the empty string instead.

Bug: T311389
Change-Id: I2bd4315638642a9f56f31922cdd31721ad212ede