Remove meaningless @package and @group tags
These tags are not consistent and cannot really be used for
anything. I believe this was all auto-generated at some point
by some IDE but doesn't do anything meaningful but repeat the
namespace. We can see this in the git history: these tags are all
added in 2016 by the same user with the original commit that added
the files.

The only possible use case is when we want to run only the tests
from a specific group (which we actively do with e.g. `@group
Database`). But there are other, more reliable ways to run the test
from one extension, e.g. filtering by path.

Change-Id: If45751f4a8851587636a59f8ab380bd9ecb419ff