Replace use of func_get_args() in JCObjContent
Use variadic argument for the extra validators (as already done in two
of the three places) and avoid func_get_args() to access the argument
list.
The extra convert function is needed to detect an array of validators
given as one argument instead of variadic args.

Gives less magic and better IDE support, as the extra validators are now
documented in all functions.

Change-Id: If675605b6690b2bee6abf2027685d27114ff47f3