Re-namespace the extension's code to MediaWiki\Extension\JsonConfig
This is the expected naming. I've slapped a class-alias on all of the
classes, as at least a few of them seem to be used in other repos and
it's easier to drop them one-by-one later than add them in a panic in
a back-port once it's already broken production.

Change-Id: I336097d18eab00480657bd5a76fcf91da10d4fea