Support for Data: page global usage display
Moved namespace/namespace text back into the grouping with the wiki,
so we can reduce duplicate strings while still retaining namespace
text to make functional and correct links. The index was tweaked per
ladsgroup's recommendation to allow for cases of multiple namespace
text variants -- this happens with gendered user namespaces in some
languages like Polish.

Eg if a Data: page is invoked on someone's User: or User_talk: page in
affected language wikis, it should record the correct form of the link
for display as GlobalUsage does -- but more efficiently without massive
string dupes.

Tests have been updated, including a test case for the updating of
legacy row data when running the links updater logic.

Adapted some bits from GlobalUsage, but stripped down a lot.
Did not include the extra machinery to run a restartable
query (for API or special page lookups), but we can add that
in the future if it's desired.

Schema change adds gjlw_namespace and gjlw_namespace_text fields to
globaljsonlinks_wiki table, which lives in x1.commonswiki in the
production system:

* sql/mysql/patch-gjlw_namespace_text.sql

On standalone systems, the updater will run on the in-place copy
of the globaljsonlinks table, but these are not precent in prod.

To keep running on the old schema, set the feature flag off:

  $wgTrackGlobalJsonLinksNamespaces = false;

and let it go back to the deafult of enabled after deploying
the schema update.

Existing rows with a NULL gjlw_namespace_text will be displayed
using canonical namespace names or a fallback numeric ID. As the
referencing pages get edited, they will update their records
automatically to include the current wiki/language/user-specific
localization.

Note this should show usage of both tabular data via the Lua loader
function and invocation of Charts both on the format and data pages.
This is tracked in the same table.

Bug: T371300
Bug: T383591
Change-Id: I4bfa50e8aceb05e362e41a1d5f486f3b0688a350