Create new service 'JsonConfig.Utils' for class JCUtils
In the second step (Ia29e834a3ce14039066da77d3718fa2d62af71a7) the
static function isLocalizedArray gets replaced by a non-static function
to use the dependency injection.

Callers of this functions can now switch from a static call to a
non-static call.

The tests in JCMapDataContentUnitTest are now in JCMapDataContentTest,
because the unit test fails with
"Premature access to service container".

Change-Id: I2a921cbec8a355f5bdac593b1f4c873ea2f9b89b