Lua transform backend for JsonConfig Data: pages
JCContentLoader and JCTransform pipeline for loading Data: JSON pages
with optional transforms by Lua functions, when fetching locally or
remote (via action=jsontransform)

Initially this is disabled to coordinate testing and rollout;
to enable Data: pipeline transforms set the config var:

  $wgJsonConfigTransformsEnabled = true;

without which it will reject all attempts to run transforms with a
clean error and disable the action=jsontransform API module.

Call JCSingleton::getContentLoader() to get a builder, optionally
pass a JCTransform setup reference in with a module/function and
some arguments, and load() to fetch the resulting success or failed
Status error.

Modified output is validated as though it were being inserted into
a fresh edit, and should report validation errors back through
the Status object fairly sensibly though some cases may bring
big exception backtraces. ;)

This is to be used initially for Charts, allowing a Data:.chart desc
to specify a transform to incoming Data:.tab on the way to rendering.

Execution remains centralized in the store wiki (Commons in production)
and doesn't expose any new internal interfaces, so should have a small
security surface area impact (Lua functions can already run on
this data via action=parse etc)

This commit sets up a test set from my 'Canadian election' sample
data set and a small sample Lua module with three transform functions,
respectively to extract a subset of columns, to extract a subset of
rows where a particula field matches a value, and to add a sum
column for a given group of input columns.

JCTransform class encapsulates the transform arguments and implements
the call into Scribunto.

JCTransformTest adds test cases to exercise this.

Pages loaded via Lua scripts will appear in the templatelinks output
of the ParserOutput, and these are exported through to `dependencies`
on the output JSON to be recorded via $wrapper->addToParserOutput()

Bug: T388434
Change-Id: I4592b93df6119ab795bb872dc38635232d9b204f