Fix for multiple charts on same page using mix of transforms
Was accidentally letting a JCContentLoader get retained in the
service container, leading to calls without transforms retaining
the most recently set transform properties instead of none.

Adds a test case for this, and a fix by using a separate factory
which is retained instead.

Also includes a fix for unit tests in JsonConfig that crops up on
my system but not on CI (??) by adding an explicit clone on an
object copy in a test setup.

Bug: T396512
Change-Id: I2ba888c6ae43115aa3062f1f81d88a24f5be0756
(cherry picked from commit 2aade05da731ecd6dba6758d397111c81e46e979)