tests: Integration tests for JCCache class
This test suite covers the 3 basic cases of JCCache::get()
^ when it returns false,
^ when it returns JCContent,
^ and when it returns string.

TODO: Expand test suite to cover more cases. Even though this
proves the failure that happen when we migrated to getWithSet().

The purpose of this is to simulate the failure caused by the patch
02f0ee3d66 (Ic2e838f6f96e0ade4) which was reverted in 9047f6cb69b5
(I74fc4502bad).

A revert revert (x2) has been made in I414c29325863acf0c22eb2b0
which will depend on this patch. The revert revert should fail until
it is fixed.

Change-Id: I83100cf3acbc4aa11fa6809f749affb30e3d4240