api: Drop support for `content` parameter with `reload` command
Since Id717fcc3bb46a9b2865 (a23b534d0a0312c) when the content param
was introduced, it seems to have never been used after that because
the check that was suppose to trigger putting the supplied content into
the cache (for the corresponding key) was checking against the incorrect
string literal "reload " (with a space).

Today, supplying a content override with the reload command doesn't
work (e.g., on `ruwiki`) where it is enabled, and on a local setup as
well.

The plain (reload) functionality is still available via the API that
can be used to store content from the DB (for a given key) to the cache.

Notes:

^ Support for the content override parameter has been removed.

^ The content override param removed from the JCCache class constructor.

^ Update tests to remove those with content overrides as they will
  now fail.

^ Remove the `WANCache->set()` call that was used to update the
  cache with a new value via the action API.

Bug: T399314
Change-Id: Idf4e84a45b5967aa0d6d999695d3c2f9eade2455