Database index hack to speed chartinfo API
There's a LOT of rows tracking usage of .tab and .map pages now, more than
there are for .chart pages, and the action=chartinfo API query's backend
here was hitting the wrong end of the database query optimization.

Forcing it to filter with the title index first chops out many untold
thousands of rows and it returns the data set from the index very quickly
now.

Note that if we want long-term breakdowns for .tab and .map this won't
be sustainable, we'll need a summary table of some kind to map suffixes.

Bug: T393950
Change-Id: Ib59afe0cadf6d6bc8f2756e2408ee24e4f43f7cc